Move substrate structures into their own module
