+++ src/node/mod.rs
@@ -11,15 +11,10 @@ use futures::channel::{mpsc, oneshot};
 use futures::{FutureExt, SinkExt, Stream, StreamExt};
 use libp2p_core::Multiaddr;
 use sc_consensus_subspace_rpc::SegmentHeaderProvider;
-use sc_executor::{WasmExecutionMethod, WasmtimeInstantiationStrategy};
-use sc_network::config::{NodeKeyConfig, Secret};
 use sc_network::network_state::NetworkState;
 use sc_network::{NetworkService, NetworkStateInfo, NetworkStatusProvider, SyncState};
-use sc_network_common::config::{MultiaddrWithPeerId, TransportConfig};
+use sc_network_common::config::MultiaddrWithPeerId;
 use sc_rpc_api::state::StateApiClient;
-use sc_service::config::{KeystoreConfig, NetworkConfiguration, OffchainWorkerConfig};
-use sc_service::{BasePath, Configuration, DatabaseSource, TracingReceiver};
-use serde::{Deserialize, Serialize};
 use sp_consensus::SyncOracle;
 use sp_core::H256;
 use subspace_core_primitives::{PieceIndexHash, SegmentIndex};
@@ -50,165 +45,22 @@ use crate::utils::DropCollection;
 
 pub mod chain_spec;
 pub mod domains;
+mod substrate;
 
-pub use builder::{
-    Base, BaseBuilder, BlocksPruning, Builder, Config, Constraints, Dsn, DsnBuilder,
-    ExecutionStrategy, Network, NetworkBuilder, OffchainWorker, OffchainWorkerBuilder, PruningMode,
-    Rpc, RpcBuilder,
-};
-pub(crate) use builder::{ImplName, ImplVersion};
+pub use builder::{Builder, Config, Dsn, DsnBuilder};
 pub use domains::{ConfigBuilder as SystemDomainBuilder, SystemDomainNode};
+pub use substrate::*;
 
 mod builder {
-    use std::net::SocketAddr;
     use std::num::NonZeroUsize;
 
     use derivative::Derivative;
     use derive_builder::Builder;
     use derive_more::{Deref, DerefMut, Display, From};
-    use sc_network::ProtocolName;
-    use sc_network_common::config::NonDefaultSetConfig;
     use serde::{Deserialize, Serialize};
 
+    use super::substrate::Base;
     use super::*;
-
-    /// Block pruning settings.
-    #[derive(
-        Debug, Clone, Copy, PartialEq, Default, Serialize, Deserialize, Eq, PartialOrd, Ord,
-    )]
-    pub enum BlocksPruning {
-        #[default]
-        /// Keep full block history, of every block that was ever imported.
-        KeepAll,
-        /// Keep full finalized block history.
-        KeepFinalized,
-        /// Keep N recent finalized blocks.
-        Some(u32),
-    }
-
-    impl From<sc_service::BlocksPruning> for BlocksPruning {
-        fn from(value: sc_service::BlocksPruning) -> Self {
-            match value {
-                sc_service::BlocksPruning::KeepAll => Self::KeepAll,
-                sc_service::BlocksPruning::KeepFinalized => Self::KeepFinalized,
-                sc_service::BlocksPruning::Some(n) => Self::Some(n),
-            }
-        }
-    }
-
-    impl From<BlocksPruning> for sc_service::BlocksPruning {
-        fn from(value: BlocksPruning) -> Self {
-            match value {
-                BlocksPruning::KeepAll => Self::KeepAll,
-                BlocksPruning::KeepFinalized => Self::KeepFinalized,
-                BlocksPruning::Some(n) => Self::Some(n),
-            }
-        }
-    }
-
-    /// Pruning constraints. If none are specified pruning is
-    #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-    pub struct Constraints {
-        /// Maximum blocks. Defaults to 0 when unspecified, effectively keeping
-        /// only non-canonical states.
-        pub max_blocks: Option<u32>,
-    }
-
-    impl From<Constraints> for sc_state_db::Constraints {
-        fn from(Constraints { max_blocks }: Constraints) -> Self {
-            Self { max_blocks }
-        }
-    }
-
-    impl From<sc_state_db::Constraints> for Constraints {
-        fn from(sc_state_db::Constraints { max_blocks }: sc_state_db::Constraints) -> Self {
-            Self { max_blocks }
-        }
-    }
-
-    /// Pruning mode.
-    #[derive(Debug, Clone, Eq, PartialEq, Default, Serialize, Deserialize)]
-    pub enum PruningMode {
-        /// No pruning. Canonicalization is a no-op.
-        #[default]
-        ArchiveAll,
-        /// Canonicalization discards non-canonical nodes. All the canonical
-        /// nodes are kept in the DB.
-        ArchiveCanonical,
-        /// Maintain a pruning window.
-        Constrained(Constraints),
-    }
-
-    impl From<PruningMode> for sc_service::PruningMode {
-        fn from(value: PruningMode) -> Self {
-            match value {
-                PruningMode::ArchiveAll => Self::ArchiveAll,
-                PruningMode::ArchiveCanonical => Self::ArchiveCanonical,
-                PruningMode::Constrained(c) => Self::Constrained(c.into()),
-            }
-        }
-    }
-
-    impl From<sc_service::PruningMode> for PruningMode {
-        fn from(value: sc_service::PruningMode) -> Self {
-            match value {
-                sc_service::PruningMode::ArchiveAll => Self::ArchiveAll,
-                sc_service::PruningMode::ArchiveCanonical => Self::ArchiveCanonical,
-                sc_service::PruningMode::Constrained(c) => Self::Constrained(c.into()),
-            }
-        }
-    }
-
-    /// Strategy for executing a call into the runtime.
-    #[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Deserialize, Serialize)]
-    pub enum ExecutionStrategy {
-        /// Execute with the native equivalent if it is compatible with the
-        /// given wasm module; otherwise fall back to the wasm.
-        #[default]
-        NativeWhenPossible,
-        /// Use the given wasm module.
-        AlwaysWasm,
-        /// Run with both the wasm and the native variant (if compatible).
-        /// Report any discrepancy as an error.
-        Both,
-        /// First native, then if that fails or is not possible, wasm.
-        NativeElseWasm,
-    }
-
-    impl From<sc_service::config::ExecutionStrategy> for ExecutionStrategy {
-        fn from(value: sc_service::config::ExecutionStrategy) -> Self {
-            use sc_service::config::ExecutionStrategy as Other;
-            match value {
-                Other::Both => Self::Both,
-                Other::AlwaysWasm => Self::AlwaysWasm,
-                Other::NativeWhenPossible => Self::NativeWhenPossible,
-                Other::NativeElseWasm => Self::NativeElseWasm,
-            }
-        }
-    }
-
-    impl From<ExecutionStrategy> for sc_service::config::ExecutionStrategy {
-        fn from(value: ExecutionStrategy) -> Self {
-            match value {
-                ExecutionStrategy::Both => Self::Both,
-                ExecutionStrategy::AlwaysWasm => Self::AlwaysWasm,
-                ExecutionStrategy::NativeWhenPossible => Self::NativeWhenPossible,
-                ExecutionStrategy::NativeElseWasm => Self::NativeElseWasm,
-            }
-        }
-    }
-
-    impl From<ExecutionStrategy> for sc_service::config::ExecutionStrategies {
-        fn from(value: ExecutionStrategy) -> Self {
-            sc_service::config::ExecutionStrategies {
-                syncing: value.into(),
-                importing: value.into(),
-                block_construction: value.into(),
-                offchain_worker: value.into(),
-                other: value.into(),
-            }
-        }
-    }
 
     #[derive(
         Debug,
@@ -305,524 +157,6 @@ mod builder {
         /// Devnet configuraiton
         pub fn devnet() -> Builder {
             Builder::devnet()
-        }
-    }
-
-    #[derive(
-        Debug,
-        Clone,
-        Derivative,
-        Deserialize,
-        Serialize,
-        PartialEq,
-        Eq,
-        From,
-        Deref,
-        DerefMut,
-        Display,
-    )]
-    #[derivative(Default)]
-    #[serde(transparent)]
-    pub struct ImplName(
-        #[derivative(Default(value = "env!(\"CARGO_PKG_NAME\").to_owned()"))] pub(crate) String,
-    );
-
-    #[derive(
-        Debug,
-        Clone,
-        Derivative,
-        Deserialize,
-        Serialize,
-        PartialEq,
-        Eq,
-        From,
-        Deref,
-        DerefMut,
-        Display,
-    )]
-    #[derivative(Default)]
-    #[serde(transparent)]
-    pub struct ImplVersion(
-        #[derivative(Default(
-            value = "format!(\"{}-{}\", env!(\"CARGO_PKG_VERSION\"), env!(\"GIT_HASH\"))"
-        ))]
-        pub(crate) String,
-    );
-
-    #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-    pub struct StorageMonitor {
-        #[serde(with = "bytesize_serde")]
-        pub threshold: bytesize::ByteSize,
-        pub polling_period: std::time::Duration,
-    }
-
-    impl From<StorageMonitor> for sc_storage_monitor::StorageMonitorParams {
-        fn from(StorageMonitor { threshold, polling_period }: StorageMonitor) -> Self {
-            Self {
-                threshold: (threshold.as_u64() / bytesize::MIB).max(1),
-                polling_period: polling_period.as_secs().max(1) as u32,
-            }
-        }
-    }
-
-    #[doc(hidden)]
-    #[derive(Debug, Clone, Derivative, Builder, Deserialize, Serialize, PartialEq)]
-    #[derivative(Default)]
-    #[builder(pattern = "immutable", build_fn(private, name = "_build"), name = "BaseBuilder")]
-    #[non_exhaustive]
-    pub struct Base {
-        /// Force block authoring
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub force_authoring: bool,
-        /// Set node role
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub role: Role,
-        /// Blocks pruning options
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub blocks_pruning: BlocksPruning,
-        /// State pruning options
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub state_pruning: PruningMode,
-        /// Set execution strategies
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub execution_strategy: ExecutionStrategy,
-        /// Implementation name
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub impl_name: ImplName,
-        /// Implementation version
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub impl_version: ImplVersion,
-        /// Rpc settings
-        #[builder(setter(into), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub rpc: Rpc,
-        /// Network settings
-        #[builder(setter(into), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub network: Network,
-        /// Offchain worker settings
-        #[builder(setter(into), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub offchain_worker: OffchainWorker,
-        /// Enable color for substrate informant
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub informant_enable_color: bool,
-        /// Additional telemetry endpoints
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub telemetry: Vec<(Multiaddr, u8)>,
-    }
-
-    #[doc(hidden)]
-    #[macro_export]
-    macro_rules! derive_base {
-        (
-            $base:ty => $builder:ident {
-                $(
-                    #[doc = $doc:literal]
-                    $field:ident : $field_ty:ty
-                ),+
-                $(,)?
-            }
-        ) => {
-            impl $builder {
-                $(
-                #[doc = $doc]
-                pub fn $field(&self, $field: impl Into<$field_ty>) -> Self {
-                    let mut me = self.clone();
-                    me.base = me.base.$field($field.into());
-                    me
-                }
-                )*
-            }
-        };
-        ( $base:ty => $builder:ident ) => {
-            $crate::derive_base!($base => $builder {
-                /// Force block authoring
-                force_authoring: bool,
-                /// Set node role
-                role: $crate::node::Role,
-                /// Blocks pruning options
-                blocks_pruning: $crate::node::BlocksPruning,
-                /// State pruning options
-                state_pruning: $crate::node::PruningMode,
-                /// Set execution strategies
-                execution_strategy: $crate::node::ExecutionStrategy,
-                /// Implementation name
-                impl_name: $crate::node::ImplName,
-                /// Implementation version
-                impl_version: $crate::node::ImplVersion,
-                /// Rpc settings
-                rpc: $crate::node::Rpc,
-                /// Network settings
-                network: $crate::node::Network,
-                /// Offchain worker settings
-                offchain_worker: $crate::node::OffchainWorker,
-                /// Enable color for substrate informant
-                informant_enable_color: bool,
-                /// Additional telemetry endpoints
-                telemetry: Vec<(libp2p_core::Multiaddr, u8)>,
-            });
-        }
-    }
-
-    impl Base {
-        pub(crate) async fn configuration<CS>(
-            self,
-            directory: impl AsRef<Path>,
-            chain_spec: CS,
-        ) -> Configuration
-        where
-            CS: sc_chain_spec::ChainSpec
-                + serde::Serialize
-                + serde::de::DeserializeOwned
-                + sp_runtime::BuildStorage
-                + 'static,
-        {
-            const NODE_KEY_ED25519_FILE: &str = "secret_ed25519";
-            const DEFAULT_NETWORK_CONFIG_PATH: &str = "network";
-
-            let Self {
-                force_authoring,
-                role,
-                blocks_pruning,
-                state_pruning,
-                execution_strategy,
-                impl_name: ImplName(impl_name),
-                impl_version: ImplVersion(impl_version),
-                rpc:
-                    Rpc {
-                        http: rpc_http,
-                        ws: rpc_ws,
-                        ws_max_connections: rpc_ws_max_connections,
-                        ipc: rpc_ipc,
-                        cors: rpc_cors,
-                        methods: rpc_methods,
-                        max_payload: rpc_max_payload,
-                        max_request_size: rpc_max_request_size,
-                        max_response_size: rpc_max_response_size,
-                        max_subs_per_conn: rpc_max_subs_per_conn,
-                        ws_max_out_buffer_capacity,
-                    },
-                network,
-                offchain_worker,
-                informant_enable_color,
-                telemetry,
-            } = self;
-
-            let base_path = BasePath::new(directory.as_ref());
-            let config_dir = base_path.config_dir(chain_spec.id());
-
-            let mut network = {
-                let builder::Network {
-                    listen_addresses,
-                    boot_nodes,
-                    force_synced,
-                    name,
-                    client_id,
-                    enable_mdns,
-                    allow_private_ip,
-                    allow_non_globals_in_dht,
-                } = network;
-                let name = name.unwrap_or_else(|| {
-                    names::Generator::with_naming(names::Name::Numbered)
-                        .next()
-                        .filter(|name| name.chars().count() < NODE_NAME_MAX_LENGTH)
-                        .expect("RNG is available on all supported platforms; qed")
-                });
-
-                let client_id = client_id.unwrap_or_else(|| format!("{impl_name}/v{impl_version}"));
-                let config_dir = config_dir.join(DEFAULT_NETWORK_CONFIG_PATH);
-                let listen_addresses = listen_addresses
-                    .into_iter()
-                    .map(|addr| {
-                        addr.to_string()
-                            .parse()
-                            .expect("Conversion between 2 libp2p versions is always right")
-                    })
-                    .collect::<Vec<_>>();
-
-                NetworkConfiguration {
-                    listen_addresses,
-                    boot_nodes: chain_spec.boot_nodes().iter().cloned().chain(boot_nodes).collect(),
-                    force_synced,
-                    transport: TransportConfig::Normal { enable_mdns, allow_private_ip },
-                    extra_sets: vec![NonDefaultSetConfig::new(
-                        ProtocolName::Static("/subspace/cross-domain-messages"),
-                        40,
-                    )],
-                    allow_non_globals_in_dht,
-                    ..NetworkConfiguration::new(
-                        name,
-                        client_id,
-                        NodeKeyConfig::Ed25519(Secret::File(
-                            config_dir.join(NODE_KEY_ED25519_FILE),
-                        )),
-                        Some(config_dir),
-                    )
-                }
-            };
-
-            // Increase default value of 25 to improve success rate of sync
-            network.default_peers_set.out_peers = 50;
-            // Full + Light clients
-            network.default_peers_set.in_peers = 25 + 100;
-            let (keystore_remote, keystore) = (None, KeystoreConfig::InMemory);
-
-            // HACK: Tricky way to add extra endpoints as we can't push into telemetry
-            // endpoints
-            let telemetry_endpoints = match chain_spec.telemetry_endpoints() {
-                Some(endpoints) => {
-                    let Ok(serde_json::Value::Array(extra_telemetry)) = serde_json::to_value(&telemetry) else {
-                        unreachable!("Will always return an array")
-                    };
-                    let Ok(serde_json::Value::Array(telemetry)) = serde_json::to_value(endpoints) else {
-                        unreachable!("Will always return an array")
-                    };
-
-                    serde_json::from_value(serde_json::Value::Array(
-                        telemetry.into_iter().chain(extra_telemetry).collect::<Vec<_>>(),
-                    ))
-                    .expect("Serialization is always valid")
-                }
-                None => sc_service::config::TelemetryEndpoints::new(
-                    telemetry.into_iter().map(|(endpoint, n)| (endpoint.to_string(), n)).collect(),
-                )
-                .expect("Never returns an error"),
-            };
-
-            Configuration {
-                impl_name,
-                impl_version,
-                tokio_handle: tokio::runtime::Handle::current(),
-                transaction_pool: Default::default(),
-                network,
-                keystore_remote,
-                keystore,
-                database: DatabaseSource::ParityDb {
-                    path: config_dir.join("paritydb").join("full"),
-                },
-                trie_cache_maximum_size: Some(67_108_864),
-                state_pruning: Some(state_pruning.into()),
-                blocks_pruning: blocks_pruning.into(),
-                wasm_method: WasmExecutionMethod::Compiled {
-                    instantiation_strategy: WasmtimeInstantiationStrategy::PoolingCopyOnWrite,
-                },
-                wasm_runtime_overrides: None,
-                execution_strategies: execution_strategy.into(),
-                rpc_http,
-                rpc_ws,
-                rpc_ipc,
-                rpc_methods: rpc_methods.into(),
-                rpc_ws_max_connections,
-                rpc_cors,
-                rpc_max_payload,
-                rpc_max_request_size,
-                rpc_max_response_size,
-                rpc_id_provider: None,
-                rpc_max_subs_per_conn,
-                ws_max_out_buffer_capacity,
-                prometheus_config: None,
-                telemetry_endpoints: Some(telemetry_endpoints),
-                default_heap_pages: None,
-                offchain_worker: offchain_worker.into(),
-                force_authoring,
-                disable_grandpa: false,
-                dev_key_seed: None,
-                tracing_targets: None,
-                tracing_receiver: TracingReceiver::Log,
-                chain_spec: Box::new(chain_spec),
-                max_runtime_instances: 8,
-                announce_block: true,
-                role: role.into(),
-                base_path: Some(base_path),
-                informant_output_format: sc_informant::OutputFormat {
-                    enable_color: informant_enable_color,
-                },
-                runtime_cache_size: 2,
-            }
-        }
-    }
-
-    #[derive(
-        Debug,
-        Clone,
-        Derivative,
-        Deserialize,
-        Serialize,
-        PartialEq,
-        Eq,
-        From,
-        Deref,
-        DerefMut,
-        Display,
-    )]
-    #[derivative(Default)]
-    #[serde(transparent)]
-    pub struct MaxSubsPerConn(#[derivative(Default(value = "1024"))] pub(crate) usize);
-
-    /// Node RPC builder
-    #[derive(Debug, Clone, Derivative, Builder, Deserialize, Serialize, PartialEq, Eq)]
-    #[derivative(Default)]
-    #[builder(pattern = "immutable", build_fn(private, name = "_build"), name = "RpcBuilder")]
-    #[non_exhaustive]
-    pub struct Rpc {
-        /// RPC over HTTP binding address. `None` if disabled.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub http: Option<SocketAddr>,
-        /// RPC over Websockets binding address. `None` if disabled.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub ws: Option<SocketAddr>,
-        /// RPC over IPC binding path. `None` if disabled.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub ipc: Option<String>,
-        /// Maximum number of connections for WebSockets RPC server. `None` if
-        /// default.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub ws_max_connections: Option<usize>,
-        /// CORS settings for HTTP & WS servers. `None` if all origins are
-        /// allowed.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub cors: Option<Vec<String>>,
-        /// RPC methods to expose (by default only a safe subset or all of
-        /// them).
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub methods: RpcMethods,
-        /// Maximum payload of rpc request/responses.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub max_payload: Option<usize>,
-        /// Maximum payload of a rpc request
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub max_request_size: Option<usize>,
-        /// Maximum payload of a rpc request
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub max_response_size: Option<usize>,
-        /// Maximum allowed subscriptions per rpc connection
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub max_subs_per_conn: Option<usize>,
-        /// Maximum size of the output buffer capacity for websocket
-        /// connections.
-        #[builder(setter(strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub ws_max_out_buffer_capacity: Option<usize>,
-    }
-
-    impl RpcBuilder {
-        /// Dev configuration
-        pub fn dev() -> Self {
-            Self::default()
-        }
-
-        /// Gemini 3d configuration
-        pub fn gemini_3d() -> Self {
-            Self::new()
-                .http("127.0.0.1:9933".parse().expect("hardcoded value is true"))
-                .ws("127.0.0.1:9944".parse().expect("hardcoded value is true"))
-                .cors(vec![
-                    "http://localhost:*".to_owned(),
-                    "http://127.0.0.1:*".to_owned(),
-                    "https://localhost:*".to_owned(),
-                    "https://127.0.0.1:*".to_owned(),
-                    "https://polkadot.js.org".to_owned(),
-                ])
-        }
-
-        /// Devnet configuration
-        pub fn devnet() -> Self {
-            Self::new()
-                .http("127.0.0.1:9933".parse().expect("hardcoded value is true"))
-                .ws("127.0.0.1:9944".parse().expect("hardcoded value is true"))
-                .cors(vec![
-                    "http://localhost:*".to_owned(),
-                    "http://127.0.0.1:*".to_owned(),
-                    "https://localhost:*".to_owned(),
-                    "https://127.0.0.1:*".to_owned(),
-                    "https://polkadot.js.org".to_owned(),
-                ])
-        }
-    }
-
-    /// Node network builder
-    #[derive(Debug, Default, Clone, Builder, Deserialize, Serialize, PartialEq)]
-    #[builder(pattern = "immutable", build_fn(private, name = "_build"), name = "NetworkBuilder")]
-    #[non_exhaustive]
-    pub struct Network {
-        /// Listen on some address for other nodes
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub enable_mdns: bool,
-        /// Listen on some address for other nodes
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub allow_private_ip: bool,
-        /// Allow non globals in network DHT
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub allow_non_globals_in_dht: bool,
-        /// Listen on some address for other nodes
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub listen_addresses: Vec<Multiaddr>,
-        /// Boot nodes
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
-        pub boot_nodes: Vec<MultiaddrWithPeerId>,
-        /// Force node to think it is synced
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub force_synced: bool,
-        /// Node name
-        #[builder(setter(into, strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub name: Option<String>,
-        /// Client id for telemetry (default is `{IMPL_NAME}/v{IMPL_VERSION}`)
-        #[builder(setter(into, strip_option), default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub client_id: Option<String>,
-    }
-
-    impl NetworkBuilder {
-        /// Dev chain configuration
-        pub fn dev() -> Self {
-            Self::default().force_synced(true).allow_private_ip(true)
-        }
-
-        /// Gemini 3d configuration
-        pub fn gemini_3d() -> Self {
-            Self::default()
-                .listen_addresses(vec![
-                    "/ip6/::/tcp/30333".parse().expect("hardcoded value is true"),
-                    "/ip4/0.0.0.0/tcp/30333".parse().expect("hardcoded value is true"),
-                ])
-                .enable_mdns(true)
-        }
-
-        /// Gemini 3d configuration
-        pub fn devnet() -> Self {
-            Self::default()
-                .listen_addresses(vec![
-                    "/ip6/::/tcp/30333".parse().expect("hardcoded value is true"),
-                    "/ip4/0.0.0.0/tcp/30333".parse().expect("hardcoded value is true"),
-                ])
-                .enable_mdns(true)
         }
     }
 
@@ -997,45 +331,6 @@ mod builder {
         }
     }
 
-    /// Offchain worker config
-    #[derive(Debug, Clone, Derivative, Builder, Deserialize, Serialize, PartialEq, Eq)]
-    #[derivative(Default)]
-    #[builder(pattern = "immutable", build_fn(name = "_build"), name = "OffchainWorkerBuilder")]
-    #[non_exhaustive]
-    pub struct OffchainWorker {
-        /// Is enabled
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub enabled: bool,
-        /// Is indexing enabled
-        #[builder(default)]
-        #[serde(default, skip_serializing_if = "crate::utils::is_default")]
-        pub indexing_enabled: bool,
-    }
-
-    impl OffchainWorkerBuilder {
-        /// Dev chain configuration
-        pub fn dev() -> Self {
-            Self::default()
-        }
-
-        /// Gemini 3d configuration
-        pub fn gemini_3d() -> Self {
-            Self::default().enabled(true)
-        }
-
-        /// Devnet configuration
-        pub fn devnet() -> Self {
-            Self::default().enabled(true)
-        }
-    }
-
-    impl From<OffchainWorker> for OffchainWorkerConfig {
-        fn from(OffchainWorker { enabled, indexing_enabled }: OffchainWorker) -> Self {
-            Self { enabled, indexing_enabled }
-        }
-    }
-
     impl Builder {
         /// Dev chain configuration
         pub fn dev() -> Self {
@@ -1088,52 +383,9 @@ mod builder {
     }
 
     crate::derive_base!(Base => Builder);
-    crate::generate_builder!(Base, Rpc, Network, Dsn, OffchainWorker);
+    crate::generate_builder!(Dsn);
 }
 
-/// Role of the local node.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub enum Role {
-    #[default]
-    /// Regular full node.
-    Full,
-    /// Actual authority.
-    Authority,
-}
-
-impl From<Role> for sc_service::Role {
-    fn from(value: Role) -> Self {
-        match value {
-            Role::Full => sc_service::Role::Full,
-            Role::Authority => sc_service::Role::Authority,
-        }
-    }
-}
-
-/// Available RPC methods.
-#[derive(Debug, Copy, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub enum RpcMethods {
-    /// Expose every RPC method only when RPC is listening on `localhost`,
-    /// otherwise serve only safe RPC methods.
-    #[default]
-    Auto,
-    /// Allow only a safe subset of RPC methods.
-    Safe,
-    /// Expose every RPC method (even potentially unsafe ones).
-    Unsafe,
-}
-
-impl From<RpcMethods> for sc_service::RpcMethods {
-    fn from(value: RpcMethods) -> Self {
-        match value {
-            RpcMethods::Auto => Self::Auto,
-            RpcMethods::Safe => Self::Safe,
-            RpcMethods::Unsafe => Self::Unsafe,
-        }
-    }
-}
-
-const NODE_NAME_MAX_LENGTH: usize = 64;
 const MAX_PROVIDER_RECORDS_LIMIT: NonZeroUsize = NonZeroUsize::new(100000).expect("100000 > 0"); // ~ 10 MB
 
 impl Config {
